fix: adapt to reth/revm API changes for PR #26565 bump
Align workspace revm/alloy-evm/reth-codecs pins and update chain
configs, handler, precompiles, metering, and RPC for revm 42.

+++ crates/common/precompiles/src/metrics.rs
@@ -154,7 +154,7 @@ impl PrecompileCallOutcome {
         Self {
             status,
             gas_used: output.gas_used,
-            state_gas_used: output.state_gas_used,
+            state_gas_used: output.state_gas_used as u64,
             gas_refunded: output.gas_refunded,
             duration_seconds,
             error,
